+++ src/main/java/net/marco01809/primesworkbench/calculators/PrimeCalculator.java
@@ -24,7 +24,7 @@ public abstract class PrimeCalculator {
 		return getName();
 	}
 
-	public int getHighestDeterminableNumber() {
+	public long getHighestDeterminableNumber() {
 		return Integer.MAX_VALUE - 1;
 	}
